+++ client/src/common/components/Modals/SheetModal.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import styled from 'styled-components/native';
 
 import {COLORS} from '../../../../../shared/src/constants/colors';
-import SETTINGS from '../../../common/constants/settings';
+import SETTINGS from '../../constants/settings';
 
 const Container = styled.View<{backgroundColor?: string}>(
   ({backgroundColor}) => ({
