+++ client/src/lib/uiLib/components/UiLibRootComponent.tsx
@@ -8,7 +8,7 @@ import * as Typography from '../../../common/components/Typography/Typography.li
 import * as UiSettings from '../../../common/constants/UiSettings.library';
 import * as Session from '../../../routes/Session/components/Session.library';
 import * as Screen from '../../../common/components/Screen/Screen.library';
-import * as Modal from '../../modal/components/Modal.library';
+import * as Modals from '../../../common/components/Modals/Modals.library';
 
 export type ComponentLibrary = {[key: string]: React.ComponentType};
 export type ComponentList = {[key: string]: Array<ComponentLibrary>};
@@ -17,7 +17,7 @@ const menuItems: ComponentList = {
   Buttons: [Buttons],
   Cards: [Cards],
   Icons: [Icons],
-  Modal: [Modal],
+  Modals: [Modals],
   Typography: [Typography],
   Screen: [Screen],
   SessionComponents: [Session],
